requirements.txt
dj-database-url==2.0.0
Django==4.0.4
django-heroku==0.3.1
gunicorn==20.1.0
openai==0.27.6
python-decouple==3.8
python-dotenv==1.0.0
django-bootstrap5